Introduce grub2-efi-binary-signed-<arch> package
As part of enabling UEFI Secure Boot, the grub2 EFI binary must be
signed with our distro key.

At the moment, the signing infrastructure isn't quite ready to perform
inline signing during package build. So to work around this, we
introduced the grub2-efi-binary-signed-<arch> packages. The purpose
of these packages is to supply a way for signed versions of the
grub efi binary to land on the end-user's filesystem.