Introduce kernel-signed-<arch> package
As part of enabling the UEFI Secure Boot chain, the kernel binary must
be signed with our distro key.

At the moment, the signing infrastructure isn't quite ready to perform
inline signing during package build. So to work around this, we
introduced the kernel-signed-<arch> packages. The purpose of these
packages is to supply a way for signed versions of the kernel binary
and the associated kernel modules to land on the end-user's
filesystem.