diskutils: add clearer error when no disk found (#716)
Our ISO installer, via diskutils, checks to see if there are valid disks
to install to before proceeding. If we boot the ISO on hardware where
the installer cannot find any disks, the installer panics with a very
cryptic error message:
    PANI[0000] unexpected end of JSON input

This message leads people to believe that the error is with their
imageconfig JSON file, but in reality, the JSON referenced here is
from the output of our lsblk command. We use lsblk to see if the
system has any disks we can install to and we get this output in JSON
format. So in the case where no supported disk is found, we end up
feeding an empty JSON input into the json.Unmarshal() and we get
this panic message.

So add a check to make sure the output from lsblk isn't empty
before we feed it to the json.Unmarshal(). Now if no supported disks
are found, you should get the following error message:
    ERRO[0000] no supported disks found
    PANI[0000] no supported disks found

Signed-off-by: Chris Co <chrco@microsoft.com>