Fix two bugs in cross gcc package (#769)
This change fixes two bugs in the cross gcc package:

Missing libmpc dependency

libmpc is needed when running the cross gcc package. A Requires: libmpc dependency was added to the cross gcc spec file to make sure that it's installed when the cross gcc package is installed.

Testing in a Mariner container: Verified by trying to install the cross gcc package without the libmpc package installed. The installation failed. I then installed the libmpc package and tried installing the cross gcc package. The installation passed and gcc was able to find libmpc.

sysroot include headers not being found by cross gcc

The cross gcc tools were unable to find the sysroot include headers when gcc was invoked. Using the -E -Wp,-v - flags when invoking gcc, I found the header include paths it was searching and saw that there was a missing symlink when comparing the output. The symlink is now created during the %post step of the cross gcc package.

Testing in a Mariner container: Verified that headers under the sysroot/include path are now found when compiling a small C++ program.