growpart: Remove flock workaround (#1007)
Regression was root caused to this upstream LTS backport cc88a819 which
refactored the BLKRRPART ioctl. This introduced behavior where this
IOCTL would not return -EBUSY anymore. The fix for this backport patch
was also recently backported (commit fc2454cc0) to restore the expected
-EBUSY behavior.

The original backport was introduced into 5.10 LTS in 5.10.20 and the
fix was introduced in 5.10.33.

We have now updated our kernel to a version that includes the fix so we
can safely remove this workaround in growpart.

Signed-off-by: Chris Co <chrco@microsoft.com>